+++ NEWS.md
@@ -4,8 +4,8 @@
 
 ### New features
 
-* New functions `as_cmdstan_mcmc()`, `as_cmdstan_mle()`, `as_cmdstan_vb()` that
-create CmdStanMCMC/MLE/VB objects directly from CmdStan CSV files. (#412)
+* New function `as_cmdstan_fit()` that creates CmdStanMCMC/MLE/VB objects
+directly from CmdStan CSV files. (#412)
 
 # cmdstanr 0.3.0